+++ geo/src/lib.rs
@@ -177,8 +177,6 @@ extern crate proj;
 extern crate rstar;
 
 pub use crate::algorithm::*;
-#[allow(deprecated)]
-pub use crate::traits::ToGeo;
 pub use crate::types::Closest;
 
 pub use geo_types::{
@@ -190,7 +188,6 @@ pub use geo_types::{
 /// This module includes all the functions of geometric calculations
 pub mod algorithm;
 mod geometry_cow;
-mod traits;
 mod types;
 mod utils;
 pub(crate) use geometry_cow::GeometryCow;